# Cache beatmapset filters