--- a/resources/views/beatmapsets/_filters.blade.php
+++ b/resources/views/beatmapsets/_filters.blade.php
@@ -1,0 +1,3 @@
+<script id="json-filters" type="application/json">
+    {!! json_encode(App\Libraries\Search\BeatmapsetSearchRequestParams::getAvailableFilters()) !!}
+</script>