--- a/resources/views/beatmapsets/index.blade.php
+++ b/resources/views/beatmapsets/index.blade.php
@@ -31,9 +31,7 @@
 @section("script")
   @parent
 
-  <script id="json-filters" type="application/json">
-    {!! json_encode($filters) !!}
-  </script>
+  {!! app('layout-cache')->getBeatmapsetFilters() !!}
 
   <script id="json-beatmaps" type="application/json">
     {!! json_encode($beatmaps) !!}