This makes restarting server required when updating language and genre but then again we don't do that, so... 🤷‍♀️

Also I think the workers are currently restarted every 500 requests (default) so it should eventually be correct.

Maybe just use this for Groups caching as well... 🤔 

Before:
```
Run 1:   1781 requests in 10.10s, 526.98MB read
Run 2:   1873 requests in 10.10s, 554.21MB read
Run 3:   1784 requests in 10.10s, 528.06MB read
Run 4:   1832 requests in 10.09s, 542.07MB read
Run 5:   1808 requests in 10.19s, 534.97MB read
Run 6:   1855 requests in 10.09s, 548.88MB read
Run 7:   1844 requests in 10.08s, 545.64MB read
Run 8:   1843 requests in 10.00s, 545.37MB read
Run 9:   1804 requests in 10.08s, 533.79MB read
Run 10:   1874 requests in 10.18s, 554.55MB read
Final run:
Running 10s test @ https://on.nanaya.pro/beatmapsets/
  4 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    65.75ms   16.69ms 408.26ms   97.58%
    Req/Sec    45.83      9.63    60.00     69.25%
  1849 requests in 10.08s, 547.10MB read
Requests/sec:    183.37
Transfer/sec:     54.26MB
```

After:
```
Run 1:   2251 requests in 10.07s, 666.38MB read
Run 2:   2179 requests in 10.16s, 644.86MB read
Run 3:   2193 requests in 10.07s, 648.90MB read
Run 4:   2253 requests in 10.08s, 666.64MB read
Run 5:   2222 requests in 10.08s, 657.47MB read
Run 6:   2224 requests in 10.08s, 658.06MB read
Run 7:   2161 requests in 10.07s, 639.42MB read
Run 8:   2244 requests in 10.08s, 663.98MB read
Run 9:   2154 requests in 10.08s, 637.35MB read
Run 10:   2210 requests in 10.15s, 653.92MB read
Final run:
Running 10s test @ https://on.nanaya.pro/beatmapsets/
  4 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    55.02ms   14.90ms 392.70ms   96.26%
    Req/Sec    54.91      7.46    70.00     88.50%
  2208 requests in 10.07s, 653.34MB read
Requests/sec:    219.28
Transfer/sec:     64.89MB
```