@notbakaneko 2022-10-05
The default max_requests for swoole is 0, so unless it's been set somewhere, it doesn't restart