@nanaya 2022-10-05
The default [here](https://github.com/laravel/octane/blob/1.x/src/Commands/StartSwooleCommand.php#L27) says 500 and it's also what I observed (pids changing during test)